Move updateBookingHoursBody interface to types.ts
